+++ src/main/kotlin/first_week_carrace/CarRace.kt
@@ -1,24 +1,24 @@
-package step4
+package first_week_carrace
 
 import java.lang.Integer.max
 
 class CarRace(
-    carsName: List<String>,
+    carNames: List<String>,
 ) {
-    private val carsStatus = carsName.map {
-        if (it.length > 5) throw IllegalArgumentException("cars name length must be less than 5")
+    private val carsStatus = carNames.map {
+        require(it.length <= 5)
         Car(it)
     }
     private var maxPosition = 1
 
-    fun getCarsPosition() = carsStatus.map { it.getCarStatus() }.toList()
+    fun getCarsPosition() = carsStatus.map { Pair(it.name, it.position) }.toList()
 
     fun doRound() {
         carsStatus.forEach { maxPosition = max(it.doForwardOrNot(), maxPosition) }
     }
 
     fun getWinners() = carsStatus.filter { car ->
-        car.getPosition() == maxPosition
+        car.position == maxPosition
     }
         .map { it.name }
         .toList()
