+++ src/main/kotlin/first_week_carrace/ForwardDecider.kt
@@ -1,0 +1,5 @@
+package first_week_carrace
+
+fun interface ForwardDecider {
+    fun execute(): Boolean
+}
