+++ src/main/kotlin/first_week_carrace/Main.kt
@@ -1,4 +1,4 @@
-package step3
+package first_week_carrace
 
 fun main() {
     CarRaceGame().startGame()
