--- src/main/kotlin/step3/ForwardDecider.kt
@@ -1,5 +1,0 @@
-package step3
-
-interface ForwardDecider {
-    fun execute(): Boolean
-}
