--- src/main/kotlin/step4/ForwardDecider.kt
@@ -1,5 +1,0 @@
-package step4
-
-interface ForwardDecider {
-    fun execute(): Boolean
-}
