--- src/main/kotlin/step4/Main.kt
@@ -1,5 +1,0 @@
-package step4
-
-fun main() {
-    CarRaceGame().startGame()
-}
