+++ src/test/kotlin/first_week_carrace/CarRaceTest.kt
@@ -1,4 +1,4 @@
-package step4
+package first_week_carrace
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
@@ -10,7 +10,7 @@ class CarRaceTest {
     fun `자동차 경주 게임 테스트`() {
         val carRace = CarRace(listOf("a", "b", "c"))
         val carsPosition = carRace.getCarsPosition()
-        assertThat(carsPosition).isEqualTo(listOf("a     : -", "b     : -", "c     : -"))
+        assertThat(carsPosition).isEqualTo(listOf(Pair("a", 1), Pair("b", 1), Pair("c", 1)))
         assertThat(carsPosition.size).isEqualTo(3)
         assertThat(carRace.getWinners()).isEqualTo(listOf("a", "b", "c"))
     }
